Create README.md
This code demonstrates a method to clean data downloaded from data.census.gov for use in GIS. This example uses TN tract-level data to extract the Total population and clean the tract GEOID field to join the table in ArcGIS.

This is the first iteration of the script. The intent is to be the logic to create a GUI program that can easily clean data for new GIS users.